+++ CHANGELOG.rst
@@ -6,6 +6,8 @@
   [Autoreload in dev mode] Allow disabling autoreload
 * `#418 <https://github.com/xitrum-framework/xitrum/issues/418>`_
   [Autoreload in dev mode] Do not reload Scala objects
+* `#426 <https://github.com/xitrum-framework/xitrum/issues/426>`_
+  Add xitrum.Config.xitrum.tmpDir
 * `#413 <https://github.com/xitrum-framework/xitrum/issues/413>`_
   [Metrics viewer] Add reconnection
 * `#425 <https://github.com/xitrum-framework/xitrum/issues/425>`_